# server/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Vello Trade API
  description: |
    REST API for the Vello Trade B2B marketplace mobile app (Expo). JSON routes are mounted under **`/api`** (the mobile client uses `baseURL: {host}/api`).

    **Socket.IO** runs on the same HTTP server. After login, connect with JWT in handshake `auth: { token: "<jwt>" }`.
    Events include `room:join`, `chat:typing`, `message:new`, `chat:room:updated`. See the repository README for details.
  version: 1.0.0
servers:
  - url: http://localhost:4000
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
paths:
...
  /api/auth/me/location:
    patch:
      tags: [Auth]
      summary: Set GeoJSON point on user for maps
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [lat, lng]
              properties:
                lat:
                  type: number
                lng:
                  type: number
...
  /api/chat/rooms:
    get:
      tags: [Chat]
      summary: My rooms
    post:
      tags: [Chat]
      summary: Open or create 1:1 room
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [withUserId]
              properties:
                withUserId:
                  type: string
  /api/chat/rooms/open:
    post:
      tags: [Chat]
      summary: Alias of POST /api/chat/rooms
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [withUserId]
              properties:
                withUserId:
                  type: string
  /api/chat/rooms/{roomId}/messages:
    get:
      tags: [Chat]
      summary: Messages in room
      parameters:
        - in: path
          name: roomId
          required: true
          schema:
            type: string
    post:
      tags: [Chat]
      summary: Send message
      parameters:
        - in: path
          name: roomId
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                body:
                  type: string
                productId:
                  type: string
                orderId:
                  type: string
...
  /api/invoices/from-order:
    post:
      tags: [Invoices]
      summary: Generate invoice from order (seller)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [orderId]
              properties:
                orderId:
                  type: string
  /api/invoices/{id}/status:
    patch:
      tags: [Invoices]
      summary: Update invoice / mark paid
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [draft, sent, paid, overdue]
                method:
                  type: string
                  enum: [cash, upi, bank, card, unknown]
                notes:
                  type: string
...
  /api/locations:
    get:
      tags: [Locations]
      summary: My saved pins
    post:
      tags: [Locations]
      summary: Add location
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [label, lat, lng]
              properties:
                label:
                  type: string
                lat:
                  type: number
                lng:
                  type: number
                address:
                  type: string
                roleContext:
                  type: string
                isPrimary:
                  type: boolean
  /api/locations/{id}:
    patch:
      tags: [Locations]
      summary: Update location
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
    delete:
      tags: [Locations]
      summary: Delete location
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
  /api/locations/{id}/primary:
    patch:
      tags: [Locations]
      summary: Set primary pin
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
...
  /api/dealer-locator/locations/{id}:
    patch:
      tags: [Locations]
      summary: Same as PATCH /api/locations/{id}
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
    delete:
      tags: [Locations]
      summary: Same as DELETE /api/locations/{id}
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
  /api/dealer-locator/locations/{id}/primary:
    patch:
      tags: [Dealer locator]
      summary: Same as PATCH /api/locations/{id}/primary
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
...